Create Class
The class having a function in Kotlin can be accessed using the class name only if the class does not have its elements. if it has its own elements then a instance of the class must be created to access the function. this concept is in the code.
Thank you