Create constant value
The const keyword is used to declare properties that are immutable in nature, i.e. read-only properties. The values are only known on the compile-time though, as a result, no values may be assigned at runtime to const variables. To be a const property, a property must meet the following requirements: 

1)Must be at the top level or a member of an object or a companion object.
2)Should start from a primitive data-type
3)No custom getter So, you can’t assign a const variable to a function or a class because the variable will be initialized at runtime rather than compile-time.